+++ lib/api.js
@@ -31,6 +31,8 @@ module.exports = settings => {
     next();
   });
 
+  app.use(require('./middleware/user'));
+
   app.use(require('./routers/user'));
 
   app.use('/establishment(s)?', require('./routers/establishment'));
