+++ lib/routers/profile/index.js
@@ -1,35 +1,13 @@
-const { omit } = require('lodash');
 const { Router } = require('express');
 const isUUID = require('uuid-validate');
-const { NotFoundError } = require('../errors');
-const { validateSchema, permissions } = require('../middleware');
+const { NotFoundError } = require('../../errors');
+const { permissions } = require('../../middleware');
 
 const router = Router({ mergeParams: true });
 
-const submit = (action) => {
-  return (req, res, next) => {
-    const params = {
-      action,
-      model: 'profile',
-      data: { ...req.body, establishment: req.establishment.id },
-      id: res.profile && res.profile.id
-    };
-    req.workflow(params)
-      .then(response => {
-        res.response = response;
-        next();
-      })
-      .catch(next);
-  };
-};
+const submit = require('./submit');
 
-const validateProfile = (req, res, next) => {
-  const ignoredFields = ['comments'];
-  return validateSchema(req.models.Profile, {
-    ...(res.profile || {}),
-    ...omit(req.body, ignoredFields)
-  })(req, res, next);
-};
+const validate = require('./validate');
 
 router.param('id', (req, res, next, id) => {
   if (!isUUID(id)) {
@@ -114,7 +92,7 @@ router.get('/:id', (req, res, next) => {
 
 router.put('/:id',
   permissions('profile.update'),
-  validateProfile,
+  validate(),
   submit('update')
 );
