+++ lib/routers/user.js
@@ -3,32 +3,8 @@ const { NotFoundError } = require('../errors');
 
 const router = Router();
 
-router.use('/', (req, res, next) => {
-  const { Profile } = req.models;
-  Promise.resolve()
-    .then(() => {
-      return Profile.query()
-        .where({ userId: req.user.id })
-        .eager('establishments');
-    })
-    .then(profiles => profiles[0])
-    .then(profile => {
-      req.profile = profile;
-    })
-    .then(() => next())
-    .catch(next);
-});
-
-router.use('/', (req, res, next) => {
-  Promise.resolve()
-    .then(() => req.user.can())
-    .then(response => response.json)
-    .then(allowedActions => {
-      req.allowedActions = allowedActions;
-    })
-    .then(() => next())
-    .catch(next);
-});
+const validate = require('./profile/validate');
+const submit = require('./profile/submit');
 
 router.get('/me', (req, res, next) => {
   if (!req.profile) {
@@ -38,5 +14,10 @@ router.get('/me', (req, res, next) => {
   res.meta.allowedActions = req.allowedActions;
   next();
 });
+
+router.put('/me', (req, res, next) => {
+  res.profile = req.profile;
+  next();
+}, validate(), submit('update'));
 
 module.exports = router;
